Add files via upload
More info here: https://www.softpedia.com/get/Others/Miscellaneous/Windows-Enabler.shtml

Basically this lets you click on some greyed out buttons in windows. Double click on the exe, and you'll see a blue rectangle in the taskbar and says On which means it's working. Sean from Tuts4you kindly patched the app, so there is no need to click on it to activate it.